Add Qwen-7B results, bf16 loader fix, all-models table and final summary figure
- Fix fp16 overflow in quantized eval loader (Qwen2.5 degenerate output); now bfloat16
- Add Qwen-2.5-7B staged truthfulness results (Original 33.3 T*I, S-PID 16.7, A-LQR 0.0; MMLU up to 37.5)
- New canonical all-models table: paper_style_table_truthfulness_all_models_methods.{csv,tex,pdf}
- Add final_summary_figure.{png,pdf}: T*I across model scales + RTP toxicity (log)
- Flag Qwen-14B rows with fp16 caveat; H-infinity at 7B/14B marked infeasible_ram
- Report updated with run-status provenance

+++ AppliedControler/eval_truthfulness_methods.py
@@ -140,14 +140,15 @@ def _load_model_for_eval(
         tokenizer.pad_token_id = tokenizer.eos_token_id
         quantization = BitsAndBytesConfig(
             load_in_4bit=True,
-            bnb_4bit_compute_dtype=torch.float16,
+            # bfloat16: Qwen2.5 overflows in float16, producing degenerate output.
+            bnb_4bit_compute_dtype=torch.bfloat16,
             bnb_4bit_quant_type="nf4",
             bnb_4bit_use_double_quant=True,
         )
         model = AutoModelForCausalLM.from_pretrained(
             model_name,
             quantization_config=quantization,
-            torch_dtype=torch.float16,
+            torch_dtype=torch.bfloat16,
             device_map="auto",
             low_cpu_mem_usage=True,
             attn_implementation="eager",
